# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "EduCRM",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    },
    {
      "url": "https://edu-crm-api-exp.onrender.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "Testimonial": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "default": "5.0"
          },
          "comment": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "student": {
            "type": "string"
          }
        }
      },
      "StudentProfile": {
        "type": "object",
        "properties": {
          "nameChinese": {
            "type": "string"
          },
          "nameEnglish": {
            "type": "string"
          },
          "profilePhoto": {
            "type": "string"
          },
          "studentContacts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "owner": {
                  "type": "string",
                  "enum": [
                    "student"
                  ]
                },
                "contactService": {
                  "type": "string",
                  "description": "The ID of the contact service"
                },
                "contactValue": {
                  "type": "string",
                  "description": "The value of the contact information"
                }
              },
              "required": [
                "owner",
                "contactService",
                "contactValue"
              ]
            }
          },
          "parentContacts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "owner": {
                  "type": "string",
                  "enum": [
                    "student"
                  ]
                },
                "contactService": {
                  "type": "string",
                  "description": "The ID of the contact service"
                },
                "contactValue": {
                  "type": "string",
                  "description": "The value of the contact information"
                }
              },
              "required": [
                "owner",
                "contactService",
                "contactValue"
              ]
            }
          },
          "school": {
            "type": "string"
          },
          "accountLogin": {
            "type": "string"
          },
          "accountEmail": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/student/request": {
      "post": {
        "tags": [
          "OneToOne"
        ],
        "summary": "Access:student.Request new one to one class",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "academicArea": {
                    "type": "string",
                    "description": "The academic area of the class."
                  },
                  "topic": {
                    "type": "string",
                    "description": "The topic of the class."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful request creation"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/student/reviews/new": {
      "post": {
        "tags": [
          "Testimonial"
        ],
        "summary": "Creates new review. Available to students.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Testimonial"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review was created"
          },
          "500": {
            "description": "Database error, while saving `Review`"
          }
        }
      }
    },
    "/student/reviews/edit": {
      "put": {
        "tags": [
          "Testimonial"
        ],
        "summary": "Edits existing review. Available to students.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Testimonial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review was updated"
          },
          "400": {
            "description": "Review doesn't exist"
          }
        }
      }
    },
    "/admin/students": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "View all students",
        "responses": {
          "200": {
            "description": "Successfully got student profiles"
          },
          "404": {
            "description": "No Data"
          }
        }
      }
    },
    "/profile/student/{profileId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "view one student profile by profile id",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully got student profile"
          },
          "404": {
            "description": "No Data"
          }
        }
      }
    },
    "/student": {
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "create new student profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
"$ref": "#/components/schemas/StudentProfile"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Can't save profile to database"
          },
          "201": {
            "description": "Successfully created profile"
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "update student profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProfile"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Profile doesn't exist"
          },
          "200": {
            "description": "Successfully updated profile"
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "summary": "Deletes profile by id",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "user": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile was deleted"
          },
          "500": {
            "description": "Can't delete profile"
          }
        }
      }
    },
...
    "/student/requests": {
      "get": {
        "tags": [
          "OneToOne"
        ],
        "summary": "Access:student. Get one-to-one class requests",
        "description": "Endpoint for students to get their one-to-one class requests.",
        "responses": {
          "200": {
            "description": "Successful request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/student/payments": {
      "get": {
        "tags": [
          "OneToOne"
        ],
        "summary": "Access:student. Get payment requests",
        "description": "Endpoint for students to get their payment requests.",
        "responses": {
          "200": {
            "description": "Successful request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/student/submit": {
      "put": {
        "tags": [
          "OneToOne"
        ],
        "summary": "Submit proof for payment",
        "description": "Endpoint for submitting proof of payment.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paymentId": {
                    "type": "string",
                    "description": "The ID of the payment request to update."
                  },
                  "photoLink": {
                    "type": "string",
                    "description": "The link to the proof of payment photo."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Link updated successfully"
          },
          "404": {
            "description": "Data not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}